Modularize updateSnapShot Complete. Detect bug when switching between two routes that use ReactHook
